requirements.txt
Django==3.1.7
flake8==3.8.4
django-environ==0.4.5
psycopg2==2.8.6
Pillow==8.1.2
django-widget-tweaks==1.4.8
django-imagekit==4.0.2